add new named constructer to seriazable rect
