+++ packages/testsweets/.vscode/settings.json
@@ -4,4 +4,5 @@
     "explorer.experimental.fileNesting.patterns": {
         "*.dart": "$(capture).g.dart, $(capture).freezed.dart, $(capture).locator.dart, $(capture).logger.dart, $(capture).router.dart",
     },
+    "dart.renameFilesWithClasses": "prompt"
 }
